refactor: clean up and organize includes and formatting in editor and renderer files
- Rearranged include statements in editor_engine_bridge.cpp for better organization.
- Simplified the getViewportTexture method implementation in EditorEngineBridge for improved readability.
- Reformatted static const char* displayModeLabels in viewport_widget.cpp for consistency.
- Standardized spacing in renderer files to enhance code clarity and maintainability.
- Changed ViewportDisplayMode enum class to use uint16_t for better memory efficiency.

+++ engine/renderer/viewport_display_mode.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <cstdint>
+
 namespace RealmEngine
 {
-    enum class ViewportDisplayMode : int
+    enum class ViewportDisplayMode : uint16_t
     {
         Lit = 0,
         Albedo,
